+++ gulpfile.js
@@ -136,7 +136,8 @@ gulp.task('copy-html', function() {
     gulp.src(
         [
             'source/tryGoal/component/shared/addRemovePluralButtons.html',
-            'source/tryGoal/component/shared/tryGoalBody.html'
+            'source/tryGoal/component/shared/tryGoalBody.html',
+            'source/tryGoal/component/shared/apiLog.html'
         ])
         .pipe(gulp.dest('dist/tryGoal/component/shared'));
     gulp.src(['source/goalList/agentGoalList.html'])