+++ source/tryGoal/tryGoal_service.js
@@ -15,6 +15,50 @@ services.factory('ConfigAPI', ['$resource', function($resource) {
     });
 }]);
 
+services.factory('agentHttpInterceptor', ['agentMemory', '$rootScope', function(agentMemory, $rootScope) {
+    
+    $rootScope.apiOutput = new Array();
+
+    return {
+
+        'request': function(config) {
+            
+            if (agentMemory.tryGoal && 
+            (config.url.endsWith('query') || config.url.endsWith('response'))) {
+                var log = {type: 'request'};
+                log.method = config.method;
+                log.url = config.url;      
+                log.params = config.params;  
+                log.data = config.data;   
+                log.headers = config.headers;    
+                $rootScope.apiOutput.push(log);
+            }
+            return config;
+        },
+    
+        'response': function(response) {
+            var log;
+            if (response.data && response.data.apiLog) {
+                log = response.data.apiLog;
+                $rootScope.apiOutput.push(log.request);
+                $rootScope.apiOutput.push(log.response);
+
+            } else if (agentMemory.tryGoal &&
+            (response.config.url.endsWith('query') || response.config.url.endsWith('response'))) {
+
+                log = {type: 'response'};
+                log.method = response.config.method;
+                log.url = response.config.url;   
+                log.data = JSON.parse(JSON.stringify(response.data)); 
+                log.headers = response.config.headers;
+                $rootScope.apiOutput.push(log);
+            }
+
+            return response;
+        }
+    };
+}]);
+
 services.factory('GoalAPI', ['$resource', 'ApiConfig', function($resource, ApiConfig) {
     return $resource('', {sessionId: '@sessionId'}, {
         startGoal: {